feat: ✨ update provider's and payment stream's info when submitting proof with mutations (#197)
* fix: :adhesive_bandage: fix compilation errors from std dependencies

* feat: :sparkles: add metadata to trie + `get_file_metadata` method in forest manager

* test: :construction: add debug statements in tests (should be removed after fix)

* feat: :sparkles: change `apply_delta` to use file metadata

* style: :rotating_light: run cargo fmt

* chore: :label: run typegen

* feat: :construction: initial structure to update providers after checkpoint challenges

* fix: :adhesive_bandage: amend review

* revert: :fire: remove unneeded info logs

* feat: :sparkles: add provider & payment stream update when deleting files using `submit_proof`

* fix: :adhesive_bandage: fix merge issues

* style: :art: remove whitespace

* test: :white_check_mark: add failure case test for update provider info after key removal

* fix: :adhesive_bandage: fix merge from main

* docs: :bulb: Add TODO in providers pallet

---------

+++ api-augment/src/interfaces/augment-api-events.ts
@@ -377,7 +377,7 @@ declare module "@polkadot/api-base/types/events" {
           fingerprint: H256,
           multiaddresses: Vec<Bytes>,
           owner: AccountId32,
-          size_: u32
+          size_: u64
         ],
         {
           bspId: H256;
@@ -386,7 +386,7 @@ declare module "@polkadot/api-base/types/events" {
           fingerprint: H256;
           multiaddresses: Vec<Bytes>;
           owner: AccountId32;
-          size_: u32;
+          size_: u64;
         }
       >;
       /**
@@ -535,7 +535,7 @@ declare module "@polkadot/api-base/types/events" {
           bucketId: H256,
           location: Bytes,
           fingerprint: H256,
-          size_: u32,
+          size_: u64,
           peerIds: Vec<Bytes>
         ],
         {
@@ -544,7 +544,7 @@ declare module "@polkadot/api-base/types/events" {
           bucketId: H256;
           location: Bytes;
           fingerprint: H256;
-          size_: u32;
+          size_: u64;
           peerIds: Vec<Bytes>;
         }
       >;
@@ -1091,8 +1091,8 @@ declare module "@polkadot/api-base/types/events" {
        **/
       DynamicRatePaymentStreamCreated: AugmentedEvent<
         ApiType,
-        [userAccount: AccountId32, providerId: H256, amountProvided: u32],
-        { userAccount: AccountId32; providerId: H256; amountProvided: u32 }
+        [userAccount: AccountId32, providerId: H256, amountProvided: u64],
+        { userAccount: AccountId32; providerId: H256; amountProvided: u64 }
       >;
       /**
        * Event emitted when a dynamic-rate payment stream is removed. Provides information about the User and Provider of the stream.
@@ -1108,8 +1108,8 @@ declare module "@polkadot/api-base/types/events" {
        **/
       DynamicRatePaymentStreamUpdated: AugmentedEvent<
         ApiType,
-        [userAccount: AccountId32, providerId: H256, newAmountProvided: u32],
-        { userAccount: AccountId32; providerId: H256; newAmountProvided: u32 }
+        [userAccount: AccountId32, providerId: H256, newAmountProvided: u64],
+        { userAccount: AccountId32; providerId: H256; newAmountProvided: u64 }
       >;
       /**
        * Event emitted when a fixed-rate payment stream is created. Provides information about the Provider and User of the stream
@@ -1544,8 +1544,8 @@ declare module "@polkadot/api-base/types/events" {
        **/
       BspRequestSignUpSuccess: AugmentedEvent<
         ApiType,
-        [who: AccountId32, multiaddresses: Vec<Bytes>, capacity: u32],
-        { who: AccountId32; multiaddresses: Vec<Bytes>; capacity: u32 }
+        [who: AccountId32, multiaddresses: Vec<Bytes>, capacity: u64],
+        { who: AccountId32; multiaddresses: Vec<Bytes>; capacity: u64 }
       >;
       /**
        * Event emitted when a Backup Storage Provider has signed off successfully. Provides information about
@@ -1558,8 +1558,8 @@ declare module "@polkadot/api-base/types/events" {
        **/
       BspSignUpSuccess: AugmentedEvent<
         ApiType,
-        [who: AccountId32, multiaddresses: Vec<Bytes>, capacity: u32],
-        { who: AccountId32; multiaddresses: Vec<Bytes>; capacity: u32 }
+        [who: AccountId32, multiaddresses: Vec<Bytes>, capacity: u64],
+        { who: AccountId32; multiaddresses: Vec<Bytes>; capacity: u64 }
       >;
       /**
        * Event emitted when a SP has changed its capacity successfully. Provides information about
@@ -1567,8 +1567,8 @@ declare module "@polkadot/api-base/types/events" {
        **/
       CapacityChanged: AugmentedEvent<
         ApiType,
-        [who: AccountId32, oldCapacity: u32, newCapacity: u32, nextBlockWhenChangeAllowed: u32],
-        { who: AccountId32; oldCapacity: u32; newCapacity: u32; nextBlockWhenChangeAllowed: u32 }
+        [who: AccountId32, oldCapacity: u64, newCapacity: u64, nextBlockWhenChangeAllowed: u32],
+        { who: AccountId32; oldCapacity: u64; newCapacity: u64; nextBlockWhenChangeAllowed: u32 }
       >;
       /**
        * Event emitted when a Main Storage Provider has requested to sign up successfully. Provides information about
@@ -1579,13 +1579,13 @@ declare module "@polkadot/api-base/types/events" {
         [
           who: AccountId32,
           multiaddresses: Vec<Bytes>,
-          capacity: u32,
+          capacity: u64,
           valueProp: PalletStorageProvidersValueProposition
         ],
         {
           who: AccountId32;
           multiaddresses: Vec<Bytes>;
-          capacity: u32;
+          capacity: u64;
           valueProp: PalletStorageProvidersValueProposition;
         }
       >;
@@ -1603,13 +1603,13 @@ declare module "@polkadot/api-base/types/events" {
         [
           who: AccountId32,
           multiaddresses: Vec<Bytes>,
-          capacity: u32,
+          capacity: u64,
           valueProp: PalletStorageProvidersValueProposition
         ],
         {
           who: AccountId32;
           multiaddresses: Vec<Bytes>;
-          capacity: u32;
+          capacity: u64;
           valueProp: PalletStorageProvidersValueProposition;
         }
       >;
